fix songs -> sounds
