+++ dojo_toolkit/dojo.py
@@ -8,7 +8,7 @@ from dojo_toolkit.code_handler import DojoCodeHandler
 from dojo_toolkit.notifier import GnomeNotifier
 from dojo_toolkit.test_runner import DoctestTestRunner
 from dojo_toolkit.timer import dojo_timer
-from dojo_toolkit.song_handler import SongHandler
+from dojo_toolkit.sound_handler import SoundHandler
 
 
 class Dojo:
@@ -17,7 +17,7 @@ class Dojo:
 
         self.round_time = round_time
 
-        self.song_player = SongHandler()
+        self.sound_player = SoundHandler()
 
         if not notifier:
             notifier = GnomeNotifier()
@@ -25,15 +25,15 @@ class Dojo:
         if not test_runner:
             test_runner = DoctestTestRunner(code_path=code_path)
 
-        self.event_handler = DojoCodeHandler(notifier=notifier, test_runner=test_runner, song_player=self.song_player)
+        self.event_handler = DojoCodeHandler(notifier=notifier, test_runner=test_runner, sound_player=self.sound_player)
         self.observer = Observer()
         self.observer.schedule(self.event_handler, code_path, recursive=False)
 
-        self.timer_thread = threading.Thread(target=dojo_timer, args=(notifier, self.round_time, self.song_player))
+        self.timer_thread = threading.Thread(target=dojo_timer, args=(notifier, self.round_time, self.sound_player))
 
     def start(self):
         self.observer.start()
-        self.song_player.play_start()
+        self.sound_player.play_start()
         self.timer_thread.start()
         try:
             while True:
