+++ dojo_toolkit/settings.py
@@ -5,4 +5,4 @@ BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 
 ASSETS_DIR = os.path.join(BASE_DIR, 'assets')
 
-SONGS_DIR = os.path.join(BASE_DIR, 'songs')
+SOUNDS_DIR = os.path.join(BASE_DIR, 'sounds')
